refactor: rootimpl should not be imported from other packages
