+++ packages/solo/fun.go
@@ -26,7 +26,6 @@ import (
 	"github.com/iotaledger/wasp/packages/vm/core/blocklog"
 	"github.com/iotaledger/wasp/packages/vm/core/governance"
 	"github.com/iotaledger/wasp/packages/vm/core/root"
-	"github.com/iotaledger/wasp/packages/vm/core/root/rootimpl"
 	"github.com/iotaledger/wasp/packages/vm/vmtypes"
 	"github.com/stretchr/testify/require"
 )
@@ -275,7 +274,7 @@ func (ch *Chain) GetInfo() (iscp.ChainID, iscp.AgentID, map[iscp.Hname]*root.Con
 	require.NoError(ch.Env.T, err)
 	require.True(ch.Env.T, ok)
 
-	contracts, err := rootimpl.DecodeContractRegistry(collections.NewMapReadOnly(res, root.VarContractRegistry))
+	contracts, err := root.DecodeContractRegistry(collections.NewMapReadOnly(res, root.VarContractRegistry))
 	require.NoError(ch.Env.T, err)
 	return chainID, chainOwnerID, contracts
 }
