+++ packages/vm/vmcontext/internal.go
@@ -8,7 +8,6 @@ import (
 	"github.com/iotaledger/wasp/packages/vm/core/blob"
 	"github.com/iotaledger/wasp/packages/vm/core/blocklog"
 	"github.com/iotaledger/wasp/packages/vm/core/root"
-	"github.com/iotaledger/wasp/packages/vm/core/root/rootimpl"
 )
 
 // creditToAccount deposits transfer from request to chain account of of the called contract
@@ -51,21 +50,21 @@ func (vmctx *VMContext) findContractByHname(contractHname iscp.Hname) (*root.Con
 	vmctx.pushCallContext(root.Contract.Hname(), nil, nil)
 	defer vmctx.popCallContext()
 
-	return rootimpl.FindContract(vmctx.State(), contractHname)
+	return root.FindContract(vmctx.State(), contractHname)
 }
 
 func (vmctx *VMContext) getChainInfo() root.ChainInfo {
 	vmctx.pushCallContext(root.Contract.Hname(), nil, nil)
 	defer vmctx.popCallContext()
 
-	return rootimpl.MustGetChainInfo(vmctx.State())
+	return root.MustGetChainInfo(vmctx.State())
 }
 
 func (vmctx *VMContext) getFeeInfo() (colored.Color, uint64, uint64) {
 	vmctx.pushCallContext(root.Contract.Hname(), nil, nil)
 	defer vmctx.popCallContext()
 
-	return rootimpl.GetFeeInfoByContractRecord(vmctx.State(), vmctx.contractRecord)
+	return root.GetFeeInfoByContractRecord(vmctx.State(), vmctx.contractRecord)
 }
 
 func (vmctx *VMContext) getBinary(programHash hashing.HashValue) (string, []byte, error) {
